fixed some formatting for the table in group page
