+++ my-app/package.json
@@ -20,5 +20,8 @@
 		"svelte": "^3.44.0",
 		"vite": "^3.1.0"
 	},
-	"type": "module"
+	"type": "module",
+	"dependencies": {
+		"svelte-smooth-corners": "^1.0.2"
+	}
 }
