+++ my-app/src/routes/send/+page.svelte
@@ -1,2 +1,2 @@
 <h1>this is where you send messages</h1>
-<a href="/create">Create a group</a>
+<a href="/group">Create a group</a>
